Remove old files and rename others
